separate .js .css, seprate ES modules, websocket dispatcher WORKS!
