+++ elephWeb/src/main/webapp/index.html
@@ -3,8 +3,7 @@
 <head>
 
 <meta http-equiv="content-type" content="text/html; charset=utf-8">
-<script type="module" src="main.js">
-</script>
+<script type="module" src="main.js"></script>
 
 </head>
